Add setup.py and reorganize code into a package
Create setup script so the program can be installed and run from the
command line with the command "gameoflife".

Because of the setup.py script, the projects code also had to be
reorganized into it's own package. Some modules was also renamed and
moved into their own sub packages to make the projects source code
more clear.

Code was also reformated with black and imports was sorted with isort.