Optimize how cell sprites are created
Before a sprite for each cell that could exist in the grid was
created. Now only cells that are visible on the grid have sprites.
Dead cells don't have sprites anymore.

All cells are instead represented in a dictionary, where a the key
the coordinates to that cell and the value is the cells state
which is represented with 0:s and 1:s. If a cell is set to int(1)
a sprite for it is created. If a cell on the otherhand is int(0)
the sprite is deleted, if there is a sprite for that cell.