Add modal to show control information
Modal is toggled with the H key and the board gets
paused if it's active when the modal is toggled and then
starts again when the modal is closed.

User can't interact with the board when the modal is active.

Can now paste patterns no matter which CTRL key is held down.